Update MatchList to support variable window height
Cells now have a fixed width and the number of cells to show in the list is
calculated from the window height. This should enable the list to be displayed
on screens of different sizes when full screen mode is enabled.

Note that the list length is only calculated when the program is started:
changing the window size whilst is is running will not update the list. At
present the window is not user resizeable so this is not a problem.