Add fullscreen support
Fullscreen mode is toggled by pressing Shift-F and changes become effective
upon the next restart.

Caution: The fullscreen setting is written to prefs.csv which requires a
different path dependent on the operating system it is running on. On OS X the
path must be relative to the directory *containing* the app bundle, whereas on
Windows it is just relative to IceTimer.exe. (Reading from prefs.csv is the
same on both systems though!)